test/requirements.txt
pytest==7.4.3
requests==2.31.0
mysql-connector-python==8.2.0 